requirements.txt
aiohttp==2.2.5
async-timeout==1.2.1
chardet==3.0.4
multidict==3.1.3
passlib==1.7.1
pony==0.7.2
psycopg2==2.7.3
PyJWT==1.5.2
yarl==0.12.0
